some changes

+++ src/lib.rs
@@ -56,11 +56,11 @@ pub type Ixs = usize;
 // indexation of tokens 
 pub type Ind = usize;
 // for tokens quantity
-pub type Quant = u64;
+pub type Quant = u32;
 // indexation of words
 pub type Ixx= usize;
 // for words quantity
-pub type Qxx = u64;
+pub type Qxx = u32;
 // Lang varints 
 pub enum Lang {
     Eng,
